Move GoldenTransformRule to the golden package
